requirements.txt
setuptools
Pillow>=11.0